move auth components to seperate folder
